Local AI Agent/requirements.txt
langchain==0.3.27
langchain-core==0.3.76
langchain-ollama==0.3.7
langgraph==0.6.7
python-dotenv==1.1.1
imap-tools==1.12.0
requests==2.32.5
